+++ app/test_with_users.py
@@ -3,11 +3,12 @@ import spotipy.util as util
 import os
 
 #Added
-import createplaylist
+import src.playlist_generation.createplaylist as createplaylist
 
 username1 = ""
 username2 = "" 
 
+# Use gettoken script to get tokens for each user you are using 
 token1 = ''
 token2 = ''
 scope = 'user-read-recently-played user-top-read user-library-modify user-library-read playlist-read-private playlist-modify-public playlist-modify-private playlist-read-collaborative'
